Add try catch to show an error message on model run part
Users kept running into an issue where they would specify the model wrong and when hitting run model they would get hit with an error. This caused them to reload the page and have to start over which was slightly annoying. Adding this try catch to display the error message and not have to reload to improve quality of life